cloud-run-subscriber/requirements.txt
flask==3.0.3
google-cloud-bigquery==3.25.0
google-cloud-pubsub==2.23.0
gunicorn==22.0.0
pytest==8.2.2
